fix: address PR #1178 review findings (F1-F4, F6-F10, F13-F17)
Merge blockers:
- F1/F3/F6: --force now only replaces runs carrying the
  _capsula/pulled.json marker; locally produced runs are never replaced
  and the already-exists errors distinguish the two cases. push (and
  push --all) refuses/skips pulled runs so lossy reconstructions cannot
  degrade the server's copy. The existence check is re-done at
  replacement time, so a directory appearing during the download is no
  longer clobbered.
- F2: the download temp directory now lives directly under the vault
  root (same filesystem, atomic rename) instead of vault/<date>/, so an
  interrupted pull cannot leave a phantom run at the depth the vault
  scanners walk.

Non-blockers taken along:
- F4 (minimum): non-ULID pull arguments fail fast with a specific error
- F7/F8: run name and vault name are validated as single path
  components before being joined into paths
- F9: reuse capsula_core::util::hex_encode instead of local copies
- F10: rename the pull-side sanitizer to checked_run_relative_path to
  stop sharing a name with the server's normalizing variant
- F13: percent-encoding no longer escapes RFC 3986 unreserved chars
- F14: shared fetch_run_files used by both run-detail handlers
- F15: verify downloaded sizes; warn on negative duration_ms; align
  duration_ms width with the server model
- F16: document the round-trip fidelity gaps in the CLI reference
- F17: regression tests for all of the above (16 integration tests)

+++ crates/capsula-api-types/src/lib.rs
@@ -141,7 +141,7 @@ pub struct RunRecord {
     pub vault: String,
     pub project_root: String,
     pub exit_code: Option<i32>,
-    pub duration_ms: Option<i64>,
+    pub duration_ms: Option<i32>,
     pub stdout: Option<String>,
     pub stderr: Option<String>,
 }
